* Merged most of the global defines & packet opcodes into globals.h. This is now shared between Aujard and Ebenezer (was painful before having to update both files, especially with things like the _USER_DATA structure!).  * Implemented a skeleton of the PUS+letter system (STORE_CLOSE should work; it will trigger LOAD_WEB_ITEMMALL).  * Implemented the WIZ_SERVER_INDEX packet.  * Implemented the saved skillbar (USERDATA_SKILLSHORTCUT) system. This is untested (but based on my old code that was fine with 1.298).  * Fixed packets being blocked because of the STATE_GAMESTART check I added earlier (did need testing, hadn't gotten to that point yet); this now relies on a flag set on character selection.  * Implemented a skeleton of the rental system handler.  * Most of the new opcodes should now be in shared/packets.h

+++ Ebenezer/SharedMem.cpp
@@ -85,7 +85,7 @@ int CSharedMemQueue::PutData(char *pBuf, int size)
 	BYTE BlockMode;
 	int index = 0, count = 0;
 
-	if( size > m_wOffset ) {
+	if( (DWORD)size > m_wOffset ) {
 		sprintf( logstr, "DataSize Over.. - %d bytes\r\n", size );
 		LogFileWrite( logstr );
 		return SMQ_PKTSIZEOVER;
